+++ tests/world/test_zone.py
@@ -1,7 +1,7 @@
 import unittest
 
 from fflogsapi.client import FFLogsClient
-from fflogsapi.constants import FIGHT_DIFFICULTY_RAID, FIGHT_DIFFICULTY_SAVAGE
+from fflogsapi.constants import FightDifficulty
 from fflogsapi.world.dataclasses import FFLogsPartition
 from fflogsapi.world.expansion import FFLogsExpansion
 from fflogsapi.world.zone import FFLogsZone
@@ -61,7 +61,7 @@ class ZoneTest(unittest.TestCase):
 
         self.assertSetEqual(
             set(retrieved_difficulties),
-            set((FIGHT_DIFFICULTY_RAID, FIGHT_DIFFICULTY_SAVAGE))
+            set((FightDifficulty.NORMAL, FightDifficulty.SAVAGE))
         )
 
     def test_encounter(self) -> None: